Media (#1)
* Add in the config fetch endpoint

Adds in code to fetch twitter's config information.

* Fix Travis complaints

* Add in media upload functionality

This change lets you upload media to twitter and check on the status of uploads that twitter is processing.

Note that unlike most of the other twitter functions, the upload function wraps three different API calls, the INIT, APPEND, and FINALIZE calls to media/upload.json. I did this instead of exposing them for clients to manage mostly because I couldn't see any case where a client would actually want to manage the individual phases of an upload.

Uploads use the base64 encoded content variant, rather than the raw file content upload variant, mostly because I couldn't get the raw file upload version to work and there was a limit to how much I was willing to fight the code to get things to work. It seems unlikely that the extra bandwidth required for the larger encoding will make any difference in any place this code is used.

* Add some exported type constants

Co-authored-by: Dan <dan@nekomaka.local>
Co-authored-by: Dan Sugalski <52919660+drswork@users.noreply.github.com>